modified:   app/Http/Controllers/Datatables/Pengadaan/PembelianList.php 	modified:   app/Http/Controllers/_02Pengadaan/PembelianController.php 	modified:   database/migrations/2024_07_11_191848_pembelian.php 	modified:   database/migrations/2024_07_11_192408_pembelian_itm.php 	modified:   database/migrations/2024_07_30_054022_barang.php 	modified:   resources/views/products/02_pengadaan/pembelian.blade.php

+++ database/migrations/2024_07_30_054022_barang.php
@@ -23,10 +23,10 @@ return new class extends Migration
             $table->string('part')->nullable();
             $table->string('mesin')->nullable();
             $table->string('satuan')->nullable();
-            $table->float('qty_permintaan', 10, 2)->nullable();
-            $table->float('qty_acc', 10, 2)->nullable();
-            $table->float('qty_diterima', 10, 2)->nullable();
-            $table->float('qty_diambil', 10, 2)->nullable();
+            $table->integer('qty_permintaan')->nullable();
+            $table->integer('qty_acc')->nullable();
+            $table->integer('qty_diterima')->nullable();
+            $table->integer('qty_diambil')->nullable();
             $table->string('pemesan')->nullable();
             $table->string('unit')->nullable();
             $table->string('peruntukan')->nullable();
